extract split page component
